# Optimize collectAncestors Performance