--- a/transaction/transaction.go
+++ b/transaction/transaction.go
@@ -518,9 +518,9 @@ func (t *Transaction) AtomicBEEF(allowPartial bool) ([]byte, error) {
 	}
 	bumps := []*MerklePath{}
 	bumpMap := map[uint32]int{}
-	txid := t.TxID().String()
-	txns := map[string]*Transaction{txid: t}
-	ancestors, err := t.collectAncestors(&txid, txns, allowPartial)
+	txid := t.TxID()
+	txns := map[string]*Transaction{txid.String(): t}
+	ancestors, err := t.collectAncestors(txid, txns, allowPartial)
 	if err != nil {
 		return nil, err
 	}