### Summary

This pull request optimizes the performance of the collectAncestors method in the transaction package by using chainhash.Hash pointers directly instead of repeatedly converting transaction IDs to strings.

Resolves #200

### Performance Improvements

**3.17x Faster Execution:**
- Old implementation: 10,276 ns/op
- New implementation: 3,246 ns/op

**78.2% Less Memory Usage:**
- Old implementation: 23,376 B/op  
- New implementation: 5,088 B/op

**74.3% Fewer Memory Allocations:**
- Old implementation: 230 allocs/op
- New implementation: 59 allocs/op

### Benchmark Results

| Test Case | Old (ns/op) | New (ns/op) | Improvement |
|-----------|-------------|-------------|-------------|
| Basic Test | 10,276 | 3,246 | 3.17x faster |
| Deep Chain (50 levels) | 129,754 | 97,532 | 1.33x faster |
| Wide Tree (50 inputs) | 30,198 | 13,650 | 2.21x faster |

### Technical Changes

1. **Modified collectAncestors signature**: Now accepts chainhash.Hash parameter to avoid redundant string conversions
2. **Optimized recursive calls**: Pass hash objects directly instead of recalculating from strings
3. **Reduced memory allocations**: Fewer intermediate string objects created during traversal

### Files Changed

- transaction/beef.go: Updated collectAncestors function implementation
- transaction/transaction.go: Modified callers to pass hash parameters
- transaction/beef_test.go: Updated tests to match new API

### Impact

This optimization significantly improves performance for transaction processing, especially beneficial for:
- Large transaction chains with many dependencies
- High-throughput transaction processing
- Memory-constrained environments

The changes maintain full backward compatibility and identical functionality while providing substantial performance gains.